requirements.txt
livekit-agents[openai,deepgram,silero]
python-dotenv
asyncio
